Seperate fetching auth token into separate service
