+++ Gemfile
@@ -7,6 +7,7 @@ gem 'bootsnap', '>= 1.1.0', require: false
 gem 'coffee-rails', '~> 4.2'
 gem 'faraday'
 gem 'jbuilder', '~> 2.8'
+gem 'jwt'
 gem 'lograge'
 gem 'logstash-event'
 gem 'puma', '~> 3.11'
