+++ app/services/allocation/api.rb
@@ -1,4 +1,3 @@
-require "byebug"
 require "json"
 require "faraday"
 
@@ -6,41 +5,32 @@ module Allocation
   class Api
     include Singleton
 
+    def initialize
+      @token = nil
+    end
+
     def fetch_status
-      byebug
       endpoint = Rails.configuration.api_host.strip + "/status"
-      token = fetch_auth_token
+      nomis_auth_token = check_auth_token
       response = Faraday.get do |req|
         req.url endpoint
-        req.headers["Authorization"] = "Bearer #{token}"
-      end 
+        req.headers["Authorization"] = "Bearer #{nomis_auth_token}"
+      end
 
       JSON.parse(response.body)
     end
 
-    # TODO: Fetching auth token should not be part of this api
+  private
 
-    def fetch_auth_token
-      response = Faraday.post do |req|
-        req.url "#{Rails.configuration.nomis_oauth_url}/auth/oauth/token?grant_type=client_credentials"
-        req.headers["Authorization"] = "Basic #{Rails.configuration.nomis_oauth_authorisation}"
-      end
-
-      response_body = JSON.parse(response.body)
-
-      Nomis::Token.new(response_body)
-    end
-  end
-end
-
-module Nomis
-  class Token
-    attr_reader :type, :expiry, :access_token
-
-    def initialize(payload)
-      @type = payload['token_type']
-      @expiry = payload['expires_in']
-      @access_token = payload['access_token']
+    def check_auth_token
+      JWT.decode(
+        @token,
+        Rails.configuration.nomis_oauth_public_key,
+        true,
+        algorithm: 'RS256'
+      )
+    rescue JWT::ExpiredSignature
+      fetch_auth_token
     end
   end
 end
